+++ euler/functions/sumdigits.cc
@@ -2,14 +2,14 @@
 #include <cmath>
 
 // Returns the sum of the digits of num.
-size_t euler::sumDigits(size_t num)
+int euler::sumDigits(size_t num)
 {
-    size_t count = 0;
+    int count = 0;
     
     while (num)
     {
         count += num % 10;
-        num = floor(num / 10);
+        num = std::floor(num / 10);
     }
     return count;
 }
